% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{delete_danglers}
\alias{delete_danglers}
\title{Try to delete stuff after the good address}
\usage{
delete_danglers(Address)
}
\arguments{
\item{Address}{:  Prefix, Name, Type.}
}
\value{
A tibble will be returned with the:
\itemize{
\item Good data (either the input or cleaned input)
\item What was removed (may be blank)
}
}
\description{
THIS IS EXPERIMENTAL AND NOT WELL TESTED
We will search for an exact match,  where there are more than 9
points with that address for direction, name, suffix
Return either the input or the cleaned value plus what was removed
}
\examples{
delete_danglers("Tulane ST #5")
}
