% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{match_exactly}
\alias{match_exactly}
\title{Search for exact matches}
\usage{
match_exactly(Street_num, Prefix = "", Street_name, Street_type = "", Zipcode)
}
\arguments{
\item{Street_num}{Number portion of address.}

\item{Prefix}{Directional prefix (N, S, E, W) if exists.}

\item{Street_name}{Name of the street.}

\item{Street_type}{Designation like RD, ST, LN, etc.}

\item{Zipcode}{The zip code for the address}
}
\value{
A tibble will be returned with the:
\itemize{
\item Longitude (Lon)
\item Latitude (Lat)
\item a success flag  (Success) (T/F)
\item Fail code describing which part of the address failed.
}
}
\description{
Search for an exact address match in the following order:
Zipcode, Street Name, Street Type, Street Number, and Direction Prefix.
On success, the Lat Long will be returned.
}
\examples{
match_exactly("1311", "", "Tulane", "ST", "77008")
}
