% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{repair_name}
\alias{repair_name}
\title{Try to repair the street name}
\usage{
repair_name(
  Street_num,
  Prefix = "",
  Street_name,
  Street_type = "",
  Zipcode,
  Distance = 2
)
}
\arguments{
\item{Street_num}{Number portion of address.}

\item{Prefix}{Directional prefix (N, S, E, W) if exists.}

\item{Street_name}{Name of the street.}

\item{Street_type}{Designation like RD, ST, LN, etc.}

\item{Zipcode}{The zip code for the address}

\item{Distance}{The maximum edit distance allowed}
}
\value{
A tibble will be returned with the:
\itemize{
\item Longitude (Lon),
\item Latitude (Lat),
\item the new name (New_name),
\item a success flag  (Success) (T/F),
\item and a Fail code describing why the repair attempt failed.

There may be several points returned if several matched.
}
}
\description{
We will search in the given zipcode for street names that are close
to the given name, using the generalized Levenshtein (edit) distance.
On success, the Lat Long will be returned. Maybe several.
We will ignore street names likely to fail badly:
\itemize{
\item names with fewer than 4 letters
\item names containing STREET (A STREET, B STREET)
\item Number streets (4TH, 5TH, etc)
\item County Roads (CR 235)
\item Farm-to-Market Roads (FM 1960)
}
}
\examples{
repair_name("1311", "", "Tuline", "ST", "77008", Distance=2)
}
