% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{repair_number}
\alias{repair_number}
\title{Try to repair the street number}
\usage{
repair_number(
  Street_num,
  Prefix = "",
  Street_name,
  Street_type = "",
  Zipcode,
  Distance = 250
)
}
\arguments{
\item{Street_num}{Number portion of address.}

\item{Prefix}{Directional prefix (N, S, E, W) if exists.}

\item{Street_name}{Name of the street.}

\item{Street_type}{Designation like RD, ST, LN, etc.}

\item{Zipcode}{The zip code for the address}

\item{Distance}{The max distance in feet to interpolate, 250 default}
}
\value{
A tibble will be returned with the:
\itemize{
\item Longitude (Lon),
\item Latitude (Lat),
\item A description of the result - how far the interpolation was (Result),
\item a success flag  (Success) (T/F),
\item and a Fail code describing why the repair attempt failed.
}
}
\description{
We will search in the given zipcode for street numbers that are available
for the given name, and then try to interpolate. We will honor odd/even
to stay on one side of the street and look no further than 3 blocks total.
On success, the Lat Long will be returned.
}
\examples{
repair_number("1313", "", "Tulane", "ST", "77008")
}
