% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{repair_prefix}
\alias{repair_prefix}
\title{Try to repair the street prefix}
\usage{
repair_prefix(Street_num, Prefix = "", Street_name, Street_type = "", Zipcode)
}
\arguments{
\item{Street_num}{Number portion of address.}

\item{Prefix}{Directional prefix (N, S, E, W) if exists.}

\item{Street_name}{Name of the street.}

\item{Street_type}{Designation like RD, ST, LN, etc.}

\item{Zipcode}{The zip code for the address}
}
\value{
A tibble will be returned with the:
\itemize{
\item Longitude (Lon),
\item Latitude (Lat),
\item The new prefix (New_prefix),
\item a success flag  (Success) (T/F),
\item and a Fail code describing why the repair attempt failed.
}
}
\description{
We will search in the given zipcode for street prefixes that are available
for the given name, and if we only find one we are done. If there are
more than one, then we will search for the exact address for each.
On success, the Lat Long will be returned.
}
\examples{
repair_prefix("1311", "", "Tulane", "", "77008")
}
