% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{repair_zipcode}
\alias{repair_zipcode}
\title{Try to repair the zipcode}
\usage{
repair_zipcode(Street_num, Prefix = "", Street_name, Street_type = "", Zipcode)
}
\arguments{
\item{Street_num}{Number portion of address.}

\item{Prefix}{Directional prefix (N, S, E, W) if exists.}

\item{Street_name}{Name of the street.}

\item{Street_type}{Designation like RD, ST, LN, etc.}

\item{Zipcode}{The zip code for the address}
}
\value{
A tibble will be returned with the:
\itemize{
\item Longitude (Lon),
\item Latitude (Lat),
\item The Zip code (New_zipcode),
\item a success flag  (Success) (T/F),
\item and a Fail code describing why the repair attempt failed.
}
}
\description{
We will search for an exact match in a zipcode, where there is no match
for the street name in the given zipcode.
On success, the Lat Long will be returned.
}
\examples{
repair_number("1311", "", "Tulane", "ST", "77088")
}
